+++ click_extra/commands.py
@@ -24,65 +24,19 @@ from __future__ import annotations
 
 from gettext import gettext as _
 import logging
-from time import perf_counter
-from typing import Any, Dict, Sequence
+from typing import Any, Dict
 
 import click
 import cloup
 
-from . import Command, Group, echo
+from . import Command, Group
 from .colorize import ExtraHelpColorsMixin
 from .parameters import ExtraOption, all_envvars, normalize_envvar, search_params
 from .colorize import ColorOption, HelpOption
 from .config import ConfigOption, ShowParamsOption
 from .logging import VerbosityOption
+from .timer import TimerOption
 from .version import VersionOption
-
-
-class TimerOption(ExtraOption):
-    """A pre-configured option that is adding a ``--time``/``--no-time`` flag to print
-    elapsed time at the end of CLI execution."""
-
-    def print_timer(self):
-        """Compute and print elapsed execution time."""
-        echo(f"Execution time: {perf_counter() - self.start_time:0.3f} seconds.")
-
-    def register_timer_on_close(self, ctx, param, value):
-        """Callback setting up all timer's machinery.
-
-        Computes and print the execution time at the end of the CLI, if option has been
-        activated.
-        """
-        # Skip timekeeping if option is not active.
-        if not value:
-            return
-
-        # Take timestamp snapshot.
-        self.start_time = perf_counter()
-
-        # Register printing at the end of execution.
-        ctx.call_on_close(self.print_timer)
-
-    def __init__(
-        self,
-        param_decls: Sequence[str] | None = None,
-        default=False,
-        expose_value=False,
-        help=_("Measure and print elapsed execution time."),
-        **kwargs,
-    ):
-        if not param_decls:
-            param_decls = ("--time/--no-time",)
-
-        kwargs.setdefault("callback", self.register_timer_on_close)
-
-        super().__init__(
-            param_decls=param_decls,
-            default=default,
-            expose_value=expose_value,
-            help=help,
-            **kwargs,
-        )
 
 
 class ExtraContext(cloup.Context):